fix: Make flat config the default
